+++ scripts/wrapper.sh
@@ -46,10 +46,10 @@ case "$1" in
       *) e "Wrapper does not support 'cargo $2'" ;;
     esac
     ;;
-  mdbook) ensure_cargo mdbook 0.4.30 ;;
-  probe-run) ensure_cargo probe-run 0.3.9 ;;
+  mdbook) ensure_cargo mdbook 0.4.34 ;;
+  probe-run) ensure_cargo probe-run 0.3.10 ;;
   rust-size) ensure_cargo cargo-binutils 0.3.6 ;;
-  taplo) ensure_cargo taplo-cli 0.8.0 ;;
+  taplo) ensure_cargo taplo-cli 0.8.1 ;;
   twiggy) ensure_cargo twiggy 0.7.0 ;;
   *) IS_CARGO=n ;;
 esac