+++ crates/runner-host/src/board/debug.rs
@@ -12,11 +12,21 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use std::sync::OnceLock;
+use std::time::Instant;
+
 use wasefire_board_api as board;
 
 pub enum Impl {}
 
 impl board::debug::Api for Impl {
+    fn time() -> u64 {
+        static ORIGIN: OnceLock<Instant> = OnceLock::new();
+        let now = Instant::now();
+        let origin = ORIGIN.get_or_init(|| now);
+        now.duration_since(*origin).as_micros() as u64
+    }
+
     fn exit(success: bool) -> ! {
         std::process::exit(if success { 0 } else { 1 })
     }