+++ examples/assemblyscript/api.ts
@@ -598,6 +598,17 @@
     len: usize,
   ): void
 
+  // Returns the time spent since some initial event.
+  //
+  // The time is in micro-seconds and may wrap before using all 64 bits. In particular,
+  // this function may constantly return zero if time is not supported.
+  @external("env", "dt")
+  export declare function debug_time(
+    // Pointer to the upper 32-bits (may be null).
+    ptr: usize,
+  // Lower 32-bits of the time.
+  ): usize
+
   // Exits the platform with an error code.
   //
   // This is used by test applets to terminate the platform and propagate the test